fix(#4): ensure the fallback directory dispatcher
Since paths like "~/all-the-icons.el" can be both a valid directory name and a
valid (text) file name, so if the user did not specify a directory dispatcher,
this path is mistakenly seen as a regular file name by the future dispatchers,
which yields an error. This commit make sure a directory name is being received
by `dirvish-preview-default-dispatcher'.


Former-commit-id: 42b7e6ee74569c528b0280eb2f1122afc28be5ac