fix(core): tweak `dirvish-reclaim`
For now, a fullframe dirvish session restores its saved window layout when
switching to a non-dirvish buffer. This is necessary for scenarios like invoking
`switch-to-buffer` or similar commands.  But if the new buffer does not
interrupt dirvish's window layout (for example place that non-dirvish buffer in
side window), then the restoration is pretty obtrusive.