Handle systems using fdfind by default
Some systems such as Debian name fd as fdfind.

This commit changes `dirvish-fd-program` do a custom variable,
indicating the user it is possible to customize its value. When
initializing the variable, it tries to look for either `fd` or
`fdfind` and set it as its default value. If neither is found, emit a
warning about `fd` not being installed.