Define new faces for all VC states (#222)
This allows themes to apply styles that are suitable for Dirvish even
when they would not necessarily look good in the modeline where the
generic VC faces are present in.  For example, the modus-themes do not
apply colours to the generic VC modeline faces, though they should
apply colour-coding in Dirvish buffers.