BREAKING: preview text file using proxy buffer (#184 #195 #270 #295)
This commit addresses:

1. make preview window read only (#270)
2. add back `dirvish-preview-setup-hook` (#184)
3. add hook `dirvish-util-buffer-init-hook` (#195)
4. prevent focus competition when preview window is focused (#295)

As of now, all of the preview buffers are just read-only "proxy" buffer (meaning
although they can be edited directly, but it's not recommended to do so since
these buffers is not fully initialized and should be only used as preview, you
can navigate through the file contents, but do not attempt to edit them in that
buffer.  More commands relates to the preview buffer might be added in the
future.